--- homie/_typing.py
@@ -1,5 +1,0 @@
-# TYPES
-from typing import (Dict)
-from .mqtt_message import MQTTMessage
-
-MessageQue = Dict[str, MQTTMessage]
